% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noop.R
\name{noop}
\alias{noop}
\title{No Operation Function}
\usage{
noop(x)
}
\arguments{
\item{x}{an object}
}
\value{
the input \code{x}
}
\description{
Function does nothing and returns input. Used for zeroth order tests
}
\examples{
noop(5)
}
